+++ src/commands/defaultcommands/disable.js
