+++ src/commands/defaultcommands/enable.js
